requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Cython==3.0.10
dj-rest-auth==6.0.0
Django==4.2.11
django-allauth==0.63.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-extensions==3.2.3
django-multiupload==0.6.1
djangorestframework==3.15.1
ecdsa==0.19.0
ez_setup==0.9
firebase-admin==6.5.0
gcloud==0.18.3
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.7
Jinja2==3.1.3
jwcrypto==1.5.6
jws==0.1.3
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==1.23.3
oauth2client==4.1.3
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodome-test-vectors==1.0.14
PyJWT==2.8.0
pyparsing==3.1.2
pyproject==1.3.1
Pyrebase4==4.7.1
python-jose==3.3.0
python-jwt==4.1.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
